# Speed up phan analysis on small projects, reduce memory usage