--- a/NEWS
+++ b/NEWS
@@ -7,6 +7,11 @@ New Features(Analysis)
   Note that `(S|T)[]` is **not** supported yet.
 + Support alternate syntax `array<T>` and `array<Key, T>` in phpdoc (PR #1213)
   Note that Phan ignores the provided value of `Key` completely right now (i.e. same as `T[]`); Key types may be supported later.
++ Speed up phan analysis on small projects, reduce memory usage (Around 0.15 seconds and 15MB)
+  This was done by deferring loading the information about internal classes and functions until that information was needed for analysis.
+
+New Features (CLI, Configs)
++ Improve default update rate of `--progress-bar` (Update it every 0.10 seconds)
 
 Bug Fixes
 + Fixes bugs in PrintfCheckerPlugin: Alignment goes before width, and objects with __toString() can cast to %s. (#1225)