… src/Phan/Language/Element/FunctionFactory.php
… src/Phan/Language/FQSEN/FullyQualifiedGlobalStructuralElement.php
… src/Phan/Language/Type/ArrayType.php
… src/Phan/Language/Type/BoolType.php
… src/Phan/Language/Type/IterableType.php
… src/Phan/Language/Type/NativeType.php
… src/Phan/Library/Map.php
… src/Phan/Ordering.php
… src/Phan/Phan.php
… src/codebase.php
… tool/make_stubs