hasClassWithFQSEN will be called when inheriting parent classes, traits,
and interfaces, so this should work with little modification.

- Daemon mode should be slightly faster (not by much)
  with user defined classes precomputed
  (Won't have to skip over as many internal classes)

Similarly, hasFunctionWithFQSEN is called on a minority of the internal
functions.

Memory usage and runtime are down on small projects(with 2 files to analyze):

- Maximum memory usage decreased from 32MB(max) to 17MB(max)
- Analysis time dropped from 0.23s to 0.07s.
  **NOTE: This is an absolute decrease** (not a percentage),
  large projects' resource usage might decrease from
  10.15s to 10s and 815M to 800M
- Unit tests are greatly sped up, due to not having to generate
  information for all of the methods and parameters of the
  internal functions and classes.

Continue to eagerly load classes and functions in daemon mode
- If there are issues creating functions and methods, know immediately.
  If analyzing a file would trigger Phan parsing many functions, that
  would increase the latency.

(Lazily loading internal constants would have negligible impact on speed,
so that wasn't done)

Unrelated: Fixes #1231 (Print progress 0.1 seconds since previous time by default)

This speeds up `tool/make_stubs`.

- Fix bug this caused when `iterable` was accessed before `array`